+++ blog-backend/blog-common/src/main/java/com/donts/dto/RoleDTO.java
@@ -16,17 +16,17 @@ import java.util.List;
 @Schema(description = "角色")
 public class RoleDTO {
 
-    @Schema(name = "roleId", description = "角色Id", type = "Integer")
-    private Integer roleId;
+    @Schema(name = "roleId", description = "角色Id", type = "Long")
+    private Long roleId;
 
     @NotBlank(message = "角色名不能为空")
     @Schema(name = "roleName", description = "角色名", type = "String")
     private String roleName;
 
     @Schema(name = "resourceIdList", description = "资源列表", type = "List<Integer>")
-    private List<Integer> resourceIds;
+    private List<Long> resourceIds;
 
     @Schema(name = "menuIdList", description = "菜单列表", type = "List<Integer>")
-    private List<Integer> menuIds;
+    private List<Long> menuIds;
 
 }
