+++ blog-backend/blog-data/src/main/java/com/donts/blog/entity/RoleMenu.java
@@ -23,17 +23,17 @@ public class RoleMenu implements Serializable {
      * 主键
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 角色id
      */
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 菜单id
      */
-    private Integer menuId;
+    private Long menuId;
 
 
 }
